Add testFailoverForMultipleNamespaces and utils
